Bugfix: two audio recorder implementations
Replaces the AudioRecorder in the device package with the one from the recorder app. Removes it from the recorder app. The device package audio recorder seems to be incomplete/outdated/unused.
